requirements.txt
mysql-connector-python
pandas
streamlit
mediapipe
opencv-python~=4.12.0.88
numpy~=2.2.6